requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Message Queue
redis==5.0.1
celery==5.3.6
kombu==5.3.4

# LLM and AI
openai==1.10.0
anthropic==0.18.0
zhipuai==2.0.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langgraph==0.0.20

# Data Processing
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0

# Configuration and Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# S3 Storage
boto3==1.34.34
minio==7.2.3

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
testcontainers==3.7.1
aiosqlite==0.19.0
pika==1.3.2

# Code Quality
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
aiofiles==23.2.1
tenacity==8.2.3
